Re-write to use a guarded and-> pipeline to process photo structs (#2)
The general gist of this refactor is to make the program "safer" and less
prone to crashing. (That said it would still be good to have a test suite.)

Image paths are now passed to a functional pipeline that runs it through 
a series of steps which try to gather data, copy it, and store a reference
in a SQLite db. If any of those steps returns `#f` the pipeline short circuits,
logs a warning, and moves on to the next file. [Note: consider cleanup?]

JSON logging:
--------------

* Adds simple macros for JSON logging to stdout
* Use an eq? symbol-hash for JSON values
* Timestamp is now in UTC and has nanoseconds
* Revert back to using current-error-port for log output

The JSON package expects an eq?-hash, so let's enforce it. Also when
converting a list to pairs cast the field value into a symbol? if
possible, and if not gensym a default one with a message.

and-> pipeline:
---------------

This greatly simplifies the program, using a guarded and-> pipeline
passing photo? records around until the files are copied and a
reference stored in the db.

* Adds guard/and-> macro to add safety to procs in and-> threading

Photo handling:
---------------

Now does a better job trying to pull a date from image files to create
the nested folder structure. Starts with exif data, then tries to parse a
date from the filename and finally falls back on `mtime` of the file. If
all else fails, we use the Unix Epoch (1/1/1970) as the date.

* Safely pulls metadata from image file to use for filenames
* Adds heuristics to generate photo date and attempts to copy
* Updates DB queries to use new photo struct
* Adds gitignore to keep compiled files out of source