Update the valid file test to account for modified time
By checking the modified time on incoming files we can ignore those
which were added to the dir before the last known processed file. This
will radically speed up incremental processing.

We still have to get the mtime of files, which is a small bottleneck,
but saves us from having to fully process files and check that they
already exist.